chore: finalize v3.22.0 stabilization and regression fixes
- Upgrade IPC protocol to v4 (u32 headers, 32MB payload limit).
- Fix GC safety violations by ensuring explicit manifest registration.
- Update vriftd to act as a proper registry authority (rkyv manifest generation).
- Complete nomenclature refactor (vrift-shim -> vrift-inception-layer) in test scripts.
- Resolve all identified regressions in demo_gc.sh and QA suite.

+++ crates/vrift-cli/src/daemon.rs
@@ -221,7 +221,7 @@ fn spawn_daemon() -> Result<()> {
     Ok(())
 }
 
-pub async fn send_request(stream: &mut UnixStream, req: VeloRequest) -> Result<u16> {
+pub async fn send_request(stream: &mut UnixStream, req: VeloRequest) -> Result<u32> {
     tracing::debug!("Sending request: {:?}", req);
     let seq_id = vrift_ipc::frame_async::send_request(stream, &req).await?;
     Ok(seq_id)
